db/requirements.txt
# Database connection
mysql-connector-python>=8.0.28
sqlalchemy>=1.4.42

# Data generation
faker>=12.0.0

# Data processing and visualization
pandas>=1.4.0
numpy>=1.22.0
openpyxl>=3.0.9
matplotlib>=3.5.1
seaborn>=0.11.2 